requirements.txt
absl-py==2.3.1
asttokens==3.0.0
astunparse==1.6.3
blinker==1.9.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.3.0
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
decorator==5.1.1
essential-generators==1.0
executing==2.1.0
Flask==3.1.2
flask-cors==6.0.1
flatbuffers==25.2.10
fonttools==4.55.3
gast==0.6.0
google-pasta==0.2.0
grpcio==1.75.0
h5py==3.14.0
idna==3.10
ipython==8.31.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
keras==3.11.3
kiwisolver==1.4.8
libclang==18.1.1
Markdown==3.9
markdown-it-py==4.0.0
MarkupSafe==3.0.2
matplotlib==3.10.0
matplotlib-inline==0.1.7
mdurl==0.1.2
ml_dtypes==0.5.3
mpmath==1.3.0
namex==0.1.0
numpy==2.2.0
opencv-python==4.10.0.84
opt_einsum==3.4.0
optree==0.17.0
packaging==24.2
parso==0.8.4
pillow==11.0.0
prompt_toolkit==3.0.48
protobuf==6.32.1
pure_eval==0.2.3
pygame==2.6.1
Pygments==2.18.0
pyparsing==3.2.1
PyPDF2==3.0.1
pypng==0.20220715.0
PyQRCode==1.2.1
python-dateutil==2.9.0.post0
requests==2.32.5
rich==14.1.0
setuptools==80.9.0
six==1.17.0
stack-data==0.6.3
sympy==1.13.3
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
termcolor==3.1.0
tk==0.1.0
traitlets==5.14.3
typing_extensions==4.15.0
urllib3==2.5.0
wcwidth==0.2.13
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.3